Issue #1382: move LiveData to ext package
